+++ demos/employees/src/main/java/is/codion/framework/demos/employees/model/EmployeeEditModel.java
@@ -65,7 +65,7 @@ public final class EmployeeEditModel extends SwingEntityEditModel {
 		//Filter the manager ComboBoxModel so that only managers from the selected department are shown,
 		//this filtering happens each time the department value is changed, either when an employee is
 		//selected or the department combo box selection changes
-		changeEvent(Employee.DEPARTMENT_FK).addDataListener(department -> {
+		valueEvent(Employee.DEPARTMENT_FK).addDataListener(department -> {
 			//only show managers from the same department as the selected employee and hide the currently
 			//selected employee to prevent an employee from being made her own manager
 			foreignKeyComboBoxModel(Employee.MANAGER_FK).includeCondition().set(manager ->